+++ src/attributecode/cmd.py
@@ -401,17 +401,17 @@ def print_config_help(ctx, param, value):
 
 
 @about.command(cls=AboutCommand,
-    short_help='Transform a CSV by applying renamings, filters and checks.')
+    short_help='Transform a CSV/JSON by applying renamings, filters and checks.')
 
 @click.argument('location',
     required=True,
-    callback=partial(validate_extensions, extensions=('.csv',)),
+    callback=partial(validate_extensions, extensions=('.csv', '.json',)),
     metavar='LOCATION',
     type=click.Path(exists=True, dir_okay=False, readable=True, resolve_path=True))
 
 @click.argument('output',
     required=True,
-    callback=partial(validate_extensions, extensions=('.csv',)),
+    callback=partial(validate_extensions, extensions=('.csv', '.json',)),
     metavar='OUTPUT',
     type=click.Path(exists=False, dir_okay=False, writable=True, resolve_path=True))
 
@@ -438,30 +438,39 @@ def print_config_help(ctx, param, value):
 
 def transform(location, output, configuration, quiet, verbose):  # NOQA
     """
-Transform the CSV file at LOCATION by applying renamings, filters and checks
-and write a new CSV to OUTPUT.
+Transform the CSV/JSON file at LOCATION by applying renamings, filters and checks
+and write a new CSV/JSON to OUTPUT.
 
-LOCATION: Path to a CSV file.
+LOCATION: Path to a CSV/JSON file.
 
-OUTPUT: Path to CSV inventory file to create.
+OUTPUT: Path to CSV/JSON inventory file to create.
     """
     from attributecode.transform import transform_csv_to_csv
+    from attributecode.transform import transform_json_to_json
     from attributecode.transform import Transformer
 
-    if not quiet:
-        print_version()
-        click.echo('Transforming CSV...')
 
     if not configuration:
         transformer = Transformer.default()
     else:
         transformer = Transformer.from_file(configuration)
 
-    errors = transform_csv_to_csv(location, output, transformer)
+    if location.endswith('.csv') and output.endswith('.csv'):
+        errors = transform_csv_to_csv(location, output, transformer)
+    elif location.endswith('.json') and output.endswith('.json'):
+        errors = transform_json_to_json(location, output, transformer)
+    else:
+        msg = 'Extension for the input and output need to be the same.'
+        click.echo(msg)
+        sys.exit()
+
+    if not quiet:
+        print_version()
+        click.echo('Transforming...')
 
     errors_count = report_errors(errors, quiet, verbose, log_file_loc=output + '-error.log')
     if not quiet and not errors:
-        msg = 'Transformed CSV written to {output}.'.format(**locals())
+        msg = 'Transformed file written to {output}.'.format(**locals())
         click.echo(msg)
     sys.exit(errors_count)